+++ app/Api/V1/Requests/DateRangeRequest.php
@@ -40,10 +40,6 @@ class DateRangeRequest extends ApiRequest
         $validator->after(
             function (Validator $validator): void {
                 if ($validator->failed()) {
-                    // set null values
-                    $this->attributes->set('start', null);
-                    $this->attributes->set('end', null);
-
                     return;
                 }
                 $start = $this->getCarbonDate('start')?->startOfDay();